Reverse the order of function overloads in typings
As an experiment, reverse the order of function overloads in TypeScript
typings from fully-applied first to fully-curried first. The goal of
this change is to improve type inference in cases where a curried
function is passed as a parameter to another function.

For example, if we wanted to write code like

    const sum = I.reduce(I.add, 0, array)

the compiler didn't accept that before, since TypeScript always chooses
the _last_ overload in cases like that. By moving the fully-applied
function to be the last overload, it works as expected.

I'm currently not aware of any downsides to this change, but time will
tell.

As a side-effect, this changes the type signatures in the README to be
the fully curried versions, but I suppose that might be a good thing,
since it might make it easier to understand which functions are curried
and which functions are not.